fix: organizando todo o projeto.
