+++ package.json
@@ -9,39 +9,40 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/metro-runtime": "~6.1.2",
+    "@expo/metro-runtime": "~55.0.9",
     "@expo/ngrok": "^4.1.0",
     "@expo/vector-icons": "^15.1.1",
     "@react-native-async-storage/async-storage": "2.2.0",
     "@react-native-clipboard/clipboard": "^1.16.3",
     "@supabase/supabase-js": "^2.99.1",
-    "babel-preset-expo": "~54.0.10",
+    "babel-preset-expo": "~55.0.8",
     "base64-arraybuffer": "^1.0.2",
-    "expo": "~54.0.33",
-    "expo-auth-session": "~7.0.10",
-    "expo-clipboard": "~8.0.8",
-    "expo-constants": "~18.0.13",
-    "expo-file-system": "~19.0.21",
-    "expo-image-picker": "~17.0.10",
-    "expo-linear-gradient": "~15.0.8",
-    "expo-linking": "~8.0.11",
-    "expo-router": "~6.0.23",
-    "expo-splash-screen": "~31.0.13",
-    "expo-status-bar": "~3.0.9",
-    "expo-web-browser": "~15.0.10",
+    "expo": "~55.0.13",
+    "expo-auth-session": "~55.0.13",
+    "expo-clipboard": "~55.0.13",
+    "expo-constants": "~55.0.13",
+    "expo-file-system": "~55.0.16",
+    "expo-image-picker": "~55.0.18",
+    "expo-linear-gradient": "~55.0.13",
+    "expo-linking": "~55.0.12",
+    "expo-router": "~55.0.11",
+    "expo-splash-screen": "~55.0.17",
+    "expo-status-bar": "~55.0.5",
+    "expo-web-browser": "~55.0.14",
+    "lottie-react-native": "~7.3.4",
     "lucide-react-native": "^0.575.0",
     "nativewind": "^4.2.2",
-    "react": "19.1.0",
-    "react-dom": "19.1.0",
-    "react-native": "0.81.5",
-    "react-native-reanimated": "~4.1.1",
+    "react": "19.2.0",
+    "react-dom": "19.2.0",
+    "react-native": "0.83.4",
+    "react-native-reanimated": "4.2.1",
     "react-native-safe-area-context": "~5.6.0",
-    "react-native-screens": "~4.16.0",
-    "react-native-svg": "15.12.1",
+    "react-native-screens": "~4.23.0",
+    "react-native-svg": "15.15.3",
     "react-native-url-polyfill": "^3.0.0",
     "react-native-vector-icons": "^10.3.0",
     "react-native-web": "^0.21.0",
-    "react-native-worklets": "^0.8.1",
+    "react-native-worklets": "0.7.2",
     "tailwindcss": "^3.3.2"
   },
   "devDependencies": {
