--- supabase/snippets/alter table sf status.sql
@@ -1,2 +1,0 @@
-ALTER TABLE sessoes_foco
-  ADD COLUMN IF NOT EXISTS status TEXT NOT NULL DEFAULT 'salvo';
